pyproject.toml
[tool.poetry]
name = "django-null"
version = "0.1.0"
description = ""
authors = ["xurshiddev555 <xurshiddev555@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.2"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
pytz = "^2024.2"
daphne = "^4.1.2"
channels = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
